Consolidated repeated random seed checking code into single location
