+++ lightworks/emulator/simulation/quick_sampler.py
@@ -14,13 +14,13 @@
 
 from collections import Counter
 from random import random
-from typing import Any, Callable
+from typing import Callable
 
 import numpy as np
 
 from ...sdk.circuit import Circuit
 from ...sdk.state import State
-from ...sdk.utils import add_heralds_to_state
+from ...sdk.utils import add_heralds_to_state, process_random_seed
 from ...sdk.utils.post_selection import PostSelectionType
 from ..backend import Backend
 from ..results import SamplingResult
@@ -207,7 +207,7 @@ class QuickSampler:
         for i, k in enumerate(pdist.keys()):
             vals[i] = k
         # Generate N random samples and then process and count output states
-        rng = np.random.default_rng(self._check_random_seed(seed))
+        rng = np.random.default_rng(process_random_seed(seed))
         samples = rng.choice(vals, p=list(pdist.values()), size=N)
         counted = dict(Counter(samples))
         return SamplingResult(counted, self.input_state)
@@ -293,12 +293,3 @@ class QuickSampler:
             pcon += p
             cdist[s] = pcon
         return cdist
-
-    def _check_random_seed(self, seed: Any) -> int | None:
-        """Process a supplied random seed."""
-        if not isinstance(seed, (int, type(None))):
-            if int(seed) == seed:
-                seed = int(seed)
-            else:
-                raise TypeError("Random seed must be an integer.")
-        return seed
