+++ lightworks/sdk/utils/matrix_utils.py
@@ -16,83 +16,7 @@
 Contains a collection of different useful functions for operations on matrices.
 """
 
-from typing import Any
-
 import numpy as np
-from scipy.stats import unitary_group
-
-
-def random_unitary(N: int, seed: int | None = None) -> np.ndarray:  # noqa: N803
-    """
-    Generate a random NxN unitary matrix. Seed can be used to produce the same
-    unitary each time the function is called.
-
-    Args:
-
-        N (int) : The dimension of the random unitary that is to be generated.
-
-        seed (int | None, optional) : Specify a random seed to repeatedly
-            produce the same unitary matrix on each function call. Defaults to
-            None, which will produce a random matrix on each call.
-
-    Returns:
-
-        np.ndarray : The created random unitary matrix.
-
-    Raises:
-
-        TypeError : In cases where the supplied random seed is not an integer
-            value.
-
-    """
-    seed = check_random_seed(seed)
-    return unitary_group.rvs(N, random_state=seed)
-
-
-def random_permutation(
-    N: int,  # noqa: N803
-    seed: int | None = None,
-) -> np.ndarray:
-    """
-    Generate a random NxN permutation. Seed can be used to produce the same
-    unitary each time the function is called.
-
-    Args:
-
-        N (int) : The dimension of the random permutation that is to be
-            generated.
-
-        seed (int | None, optional) : Specify a random seed to repeatedly
-            produce the same unitary matrix on each function call. Defaults to
-            None, which will produce a random matrix on each call.
-
-    Returns:
-
-        np.ndarray : The created random permutation matrix.
-
-    Raises:
-
-        TypeError : In cases where the supplied random seed is not an integer
-            value.
-
-    """
-    seed = check_random_seed(seed)
-    rng = np.random.default_rng(seed)
-    return rng.permutation(np.identity(N, dtype=complex))
-
-
-def check_random_seed(seed: Any) -> int | None:
-    """Process a supplied random seed."""
-    if not isinstance(seed, (int, type(None))):
-        try:
-            int(seed)
-        except Exception as e:
-            raise TypeError("Random seed must be an integer.") from e
-        if int(seed) == seed:
-            seed = int(seed)
-        else:
-            raise TypeError("Random seed must be an integer.")
-    return seed
 
 
 def check_unitary(
